Create Motodex
🏍️motoDEX is a project that is developing by leaps and bounds. The game has made a splash in the NFT market due to the following obvious advantages:

Just select a character and start the easy-to-follow gameplay. A simple plot combined with the drive of high-speed racing and colorful locations will attract fans of the classic genre of racing simulators to their screens.

Protecting financial transactions using smart contacts. The algorithm automatically transmits, tracks and updates data on funds that are pending with players. Refusal of the human factor when using modern technologies guarantees honesty and transparency of payments. Fair discrimination. Automatic price increases for new characters provide innovative motoDEX buyers with smart benefits at the contract level. The project guarantees investors who entered the game at the start of the project an advantage in monetizing their game progress.

A gaming economy that stimulates rising prices for NFT objects. The algorithm’s condition for increasing the cost of tracks and characters after each new sale ensures a stable increase in the cost of in-game items.

High liquidity with the possibility of obtaining excess profits. At the start of the project, players have a chance to increase their investment in the gameplay.

https://openbisea.io/motodex. 
https://motodex.openbisea.com/
https://www.youtube.com/watch?v=wA4iRXsjsIg.